Fixing issue #249 duplication of tags via manual entry into browser url, adding tags that already exist via the browser url field won't get added by the search controller to the tags array (using array.uniq!, damn ruby is nice :smile:).
Styled pagination link to be a button so it was obvious it was clickable (this is for when only one search result is returned and the link isn't low enough to automatically trigger a next page ajax request.

Stopped infinite scroll search suburbs broadening from triggering if there were no suburbs. This shouldn't have happend. My bad.