Fixes for issue #265 with the add-criteria-button now working on mobiles. This is because Semantic-UI's CategorySelect dropdown doesn't work on Mobiles (even on their own page). As such, there's now a  file that is a single menu (not multi-level menu) with headers to spearate sections. This menu is also searchable.
It is only shown on mobiles and tablet viewports, with the other one shown on computers. It is also scrollable and I've increased the menu height to 300px.

This includes some styling/sass tweaks to make it look like the old one. Thankfully all the JS works without modification :smile: